게시물 삭제시 권한체크

+++ src/main/java/com/kss/exam/demo/controller/UsrArticleController.java
@@ -69,12 +69,26 @@ public class UsrArticleController {
 	
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
-	public ResultData<Integer> doDelete(int id) {
+	public ResultData<Integer> doDelete(HttpSession httpSession, int id) {
+		boolean isLogined = false;
+		int memberId = -1;
 		
 		Article article = articleService.getArticleById(id);
 		
 		if(article == null) {
 			return ResultData.from("F-1", Ut.f("%d번 게시물은 존재하지 않습니다.", id));
+		}
+		
+		if(httpSession.getAttribute("loginedMemberId") != null) {
+			isLogined = true;
+			memberId = (int)httpSession.getAttribute("loginedMemberId");
+		}
+		if(isLogined == false) {
+			return ResultData.from("F-A", "로그인 되어있지 않습니다.");
+		}
+		
+		if(article.getMemberId() != memberId) {
+			return ResultData.from("F-2", "권한이 없습니다.");
 		}
 		
 		articleService.deleteArticle(id);
